feat(backend): add group endpoints & required fields

+++ backend/src/controllers/groupsController.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from "express";
 import db from "../connection";
-import { UserRow, VoteRow } from "../interfaces/models";
+import { GroupRow, UserRow, VoteRow } from "../interfaces/models";
 
 // Group Details
 export const getGroups = async (req: Request, res: Response) => {
@@ -18,24 +18,74 @@ export const getGroups = async (req: Request, res: Response) => {
 
 export const getGroup = async (req: Request, res: Response) => {
     const { id } = req.params;
+    const { from_id } = req.body;
 
-    const query =
+    if (!from_id || !id) {
+        res.status(400).json({
+            error: "All fields are required.",
+        });
+        return;
+    }
+
+    const infoQuery =
         "SELECT * FROM groups INNER JOIN questions ON groups.id = questions.group_id WHERE groups.id = ? AND questions.date = (SELECT MAX(date) FROM questions WHERE group_id = ?)";
 
-    await db
-        .query(query, [id, id])
-        .then((result) => {
-            res.status(200).json(result[0]);
-            return;
-        })
+    const votedQuery =
+        "SELECT * FROM votes INNER JOIN questions ON votes.question_id = questions.id WHERE votes.from_id = ? AND questions.group_id = ? AND questions.date = (SELECT MAX(date) FROM questions WHERE group_id = ?)";
+
+    const info = await db
+        .query<GroupRow[]>(infoQuery, [id, id])
         .catch((err) => {
             res.status(400).json({ error: err });
             return;
         });
+
+    const voted = await db
+        .query<VoteRow[]>(votedQuery, [from_id, id, id])
+        .catch((err) => {
+            res.status(400).json({ error: err });
+            return;
+        });
+
+    if (!info || !voted) {
+        res.status(400).json({
+            error: "There was an error while getting the statistics.",
+        });
+        return;
+    }
+
+    try {
+        const hasVoted = voted[0].length > 0 ? true : false;
+
+        const infoWithVoted = [
+            {
+                ...info[0][0],
+                hasVoted: hasVoted,
+            },
+        ];
+
+        res.status(200).json(infoWithVoted);
+        return;
+    } catch (err: unknown) {
+        if (err instanceof Error) {
+            res.status(400).json({ error: err.message });
+            return;
+        } else {
+            res.status(400).json({ error: "Internal server error." });
+            return;
+        }
+    }
 };
 
 export const createGroup = async (req: Request, res: Response) => {
     const { name, owner } = req.body;
+
+    if (!name || !owner) {
+        res.status(400).json({
+            error: "All fields are required.",
+        });
+        return;
+    }
 
     const query = "INSERT INTO groups VALUES (NULL, ?, ?)";
 
@@ -54,7 +104,14 @@ export const updateGroup = async (req: Request, res: Response) => {
     const { name, owner } = req.body;
     const { id } = req.params;
 
-    const query = "UPDATE groups SET name = ? SET owner_id = ? WHERE id = ?";
+    if (!name || !owner || !id) {
+        res.status(400).json({
+            error: "All fields are required.",
+        });
+        return;
+    }
+
+    const query = "UPDATE groups SET name = ?, owner_id = ? WHERE id = ?";
 
     await db
         .query(query, [name, owner, id])
@@ -69,6 +126,13 @@ export const updateGroup = async (req: Request, res: Response) => {
 
 export const removeGroup = async (req: Request, res: Response) => {
     const { id } = req.params;
+
+    if (!id) {
+        res.status(400).json({
+            error: "All fields are required.",
+        });
+        return;
+    }
 
     const query = "DELETE FROM groups WHERE id = ?";
 
@@ -88,6 +152,13 @@ export const removeGroup = async (req: Request, res: Response) => {
 export const getUsers = async (req: Request, res: Response) => {
     const { id } = req.params;
 
+    if (!id) {
+        res.status(400).json({
+            error: "All fields are required.",
+        });
+        return;
+    }
+
     const usersQuery =
         "SELECT users.id, user_id, group_id, name, streak, username FROM memberships INNER JOIN users ON users.id = memberships.user_id WHERE memberships.group_id = ?";
 
@@ -106,14 +177,7 @@ export const getUsers = async (req: Request, res: Response) => {
             return;
         });
 
-    if (!users) {
-        res.status(400).json({
-            error: "There was an error while getting the statistics.",
-        });
-        return;
-    }
-
-    if (!votes) {
+    if (!users || !votes) {
         res.status(400).json({
             error: "There was an error while getting the statistics.",
         });
@@ -152,6 +216,13 @@ export const getUsers = async (req: Request, res: Response) => {
 export const removeUser = async (req: Request, res: Response) => {
     const { user_id, group_id } = req.body;
 
+    if (!user_id || !group_id) {
+        res.status(400).json({
+            error: "All fields are required.",
+        });
+        return;
+    }
+
     const query = "DELETE FROM memberships WHERE user_id = ? AND group_id = ?";
 
     await db
@@ -169,6 +240,13 @@ export const removeUser = async (req: Request, res: Response) => {
 export const getQuestion = async (req: Request, res: Response) => {
     const { date } = req.body;
     const { id } = req.params;
+
+    if (!date || !id) {
+        res.status(400).json({
+            error: "All fields are required.",
+        });
+        return;
+    }
 
     const query = "SELECT * FROM questions WHERE group_id = ? AND date = ?";
