+++ backend/src/controllers/votesController.ts
@@ -1,9 +1,17 @@
 import { Request, Response } from "express";
 import db from "../connection";
+import { RowDataPacket } from "mysql2";
 
 // Vote Details
 export const getVotes = async (req: Request, res: Response) => {
     const { qid } = req.params;
+
+    if (!qid) {
+        res.status(400).json({
+            error: "All fields are required.",
+        });
+        return;
+    }
 
     const query = "SELECT * FROM votes WHERE question_id = ?";
 
@@ -20,12 +28,51 @@ export const getVotes = async (req: Request, res: Response) => {
 };
 
 export const submitVote = async (req: Request, res: Response) => {
-    const { from_id, to_id, question_id } = req.body;
+    const { from_id, to_id, group_id } = req.body;
 
-    const query = "INSERT INTO votes VALUES (NULL, ?, ?, ?)";
+    if (!from_id || !to_id || !group_id) {
+        res.status(400).json({
+            error: "All fields are required.",
+        });
+        return;
+    }
+
+    const questionQuery =
+        "SELECT id FROM questions WHERE group_id = ? AND date = (SELECT MAX(date) FROM questions WHERE group_id = ?)";
+
+    const insertQuery = "INSERT INTO votes VALUES (NULL, ?, ?, ?)";
+
+    const question = await db
+        .query<RowDataPacket[]>(questionQuery, [group_id, group_id])
+        .catch((err) => {
+            res.status(400).json({ error: err });
+        });
+
+    if (!question) {
+        res.status(400).json({
+            error: "There was an error while getting the question.",
+        });
+        return;
+    }
+
+    console.log(question[0]);
+
+    let question_id: number;
+
+    try {
+        question_id = question[0][0].id;
+    } catch (err: unknown) {
+        if (err instanceof Error) {
+            res.status(400).json({ error: err.message });
+            return;
+        } else {
+            res.status(400).json({ error: "Internal server error." });
+            return;
+        }
+    }
 
     await db
-        .query(query, [from_id, to_id, question_id])
+        .query(insertQuery, [from_id, to_id, question_id])
         .then((result) => {
             res.status(200).json(result[0]);
             return;
@@ -38,6 +85,13 @@ export const submitVote = async (req: Request, res: Response) => {
 export const updateVote = async (req: Request, res: Response) => {
     const { from_id, to_id, question_id } = req.body;
     const { id } = req.params;
+
+    if (!from_id || !to_id || !question_id) {
+        res.status(400).json({
+            error: "All fields are required.",
+        });
+        return;
+    }
 
     const query =
         "UPDATE votes SET from_id = ?, to_id = ?, question_id = ? WHERE id = ?";
@@ -55,6 +109,13 @@ export const updateVote = async (req: Request, res: Response) => {
 
 export const removeVote = async (req: Request, res: Response) => {
     const { id } = req.params;
+
+    if (!id) {
+        res.status(400).json({
+            error: "All fields are required.",
+        });
+        return;
+    }
 
     const query = "DELETE FROM votes WHERE id = ?";
